cleaned up unnecessary files added new one to clarify form.
using CDN instead.
cleaning up and clarifying index.html
Added notes, also i call bs on documentation being unprofessional, in a
language like JS that is inherently unclear it’s necessary.  I’d
normally put this in a wiki, but time and doubt that anyone will read
it is at hand.
changed some of the widths and added padding, but now it’s optimized
for firefox. we have yet to get to cross platformic programming so…