updated to show how to do multiple bolean values.
also code clean up, added a couple options to show I understand the concept of each form of logic. This could be expanded using this same logic/form though at that point it's just typing. I could refactor this with new stuff, but that isn't in the scope of this assignment.